less warnings, infos in clean mode

+++ src/lib/project/features/copy-manager/copy-manager-organization.backend.ts
@@ -672,7 +672,9 @@ export class CopyManagerOrganization extends CopyManagerStandalone {
         specyficFileRelativePath
       ));
       // Helpers.log(`Eqal content with temp proj: ${}`)
-      Helpers.copyFile(readyToCopyFileInLocalTempProj, destinationFilePath);
+      if (Helpers.exists(readyToCopyFileInLocalTempProj)) { // TODO QUICK_FIX when creating empty file
+        Helpers.copyFile(readyToCopyFileInLocalTempProj, destinationFilePath);
+      }
       return;
     }
